Separate all tests.
