+++ cmd/kfulint/testdata/builtin-shadow/positive_tests.go
@@ -1,17 +1,5 @@
 package checker_test
 
-func noWarnigs() {
-	var foo struct {
-		len int
-		cap int
-	}
-
-	foo.len = 123
-	foo.cap = 321
-
-	foo.len, foo.cap = foo.cap, foo.len
-}
-
 func assigningToBuiltinIdentifiers() {
 	// Types:
 	/// assigning to predeclared identifier: bool
