+++ cmd/kfulint/testdata/comments/negative_tests.go
@@ -1,0 +1,6 @@
+package checker_test
+
+// Bar returns string.
+func Bar() {
+
+}
