+++ cmd/kfulint/testdata/param-duplication/negative_tests.go
@@ -1,0 +1,10 @@
+package checker_test
+
+func good1(a, b int)                {}
+func good2(a, b int, c int32)       {}
+func good3(a, b, c int)             {}
+func good4(a int, b int32, c int64) {}
+func good5(a int)                   {}
+func good6()                        {}
+
+func mixedTypes(a, b int, c, d int64) {}
