+++ cmd/kfulint/testdata/underef/positive_tests.go
@@ -11,9 +11,6 @@ type sampleInterface interface {
 	method()
 }
 
-func (s *sampleStruct) method() {
-}
-
 func sampleCase() {
 	var k *sampleStruct
 	/// could simplify (*k).field to k.field
@@ -26,35 +23,9 @@ func sampleCase() {
 	(*k).nestedStruct.nestedField = 6
 }
 
-func sampleCase2(k sampleInterface) {
-	k.(*sampleStruct).method()
-}
-
 func sampleCase3() {
 	var k *[5]int
 
 	/// could simplify (*k)[2] to k[2]
 	(*k)[2] = 3
-}
-
-func multipleIndir1() {
-	type point struct{ x, y int }
-	pt := point{}
-
-	pt1 := &pt
-	pt2 := &pt1
-	pt3 := &pt2
-	pt4 := &pt3
-
-	_ = (*pt2).x
-
-	_ = (**pt3).x
-
-	_ = (***pt4).x
-}
-
-type myString string
-
-func convertPtr(x string) *myString {
-	return (*myString)(&x)
 }
