+++ cmd/kfulint/testdata/unexported-call/negative_tests.go
@@ -1,0 +1,6 @@
+package checker_test
+
+func (f foo) unexported() int { return 1 }
+func (f foo) Exported() int   { return 1 }
+
+func getFoo() foo { return foo{} }
