+++ cmd/kfulint/testdata/unslice/negative_tests.go
@@ -1,0 +1,9 @@
+package checker_test
+
+func noWarning() {
+	var s string
+
+	_ = s[1:]
+	_ = s[:1]
+	_ = s
+}
